Folder structure changes + minor fix to isPlaying
