+++ .gitignore
@@ -30,3 +30,7 @@
 *.exe
 *.out
 *.app
+cmake-build-debug/*
+cmake-build-debug
+.idea/*
+.idea
