+++ headers/MagicClass.h
