+++ headers/RangedClass.h
