+++ sources/RangedClass.cpp
@@ -1,10 +1,11 @@
 //
 // Created by ausro on 7/12/2018.
 //
-#include <math.h>
-#include "MeleeClass.h"
 
-MeleeClass::MeleeClass() {
+#include "../headers/RangedClass.h"
+#include <math.h>
+
+RangedClass::RangedClass() {
     this -> name = "";
     this -> healthMin = 0;
     this -> healthMax = 0;
@@ -20,11 +21,11 @@ MeleeClass::MeleeClass() {
     this -> selected = false;
 }
 
-MeleeClass::~MeleeClass() {
+RangedClass::~RangedClass() {
 
 }
 
-void MeleeClass::Initialize(std::string name) {
+void RangedClass::Initialize(std::string name) {
     this -> name = name;
     this -> healthMin = 10;
     this -> healthMax = 100;
@@ -33,7 +34,7 @@ void MeleeClass::Initialize(std::string name) {
     this -> damageMin = 2;
     this -> damageMax = 4;
     this -> manaMin = 10;
-    this -> manaMax = 200;
+    this -> manaMax = 100;
     this -> level = 1;
     this -> exp = 0;
     this -> expNext =
@@ -43,9 +44,9 @@ void MeleeClass::Initialize(std::string name) {
     this -> selected = true;
 }
 
-void MeleeClass::PrintStats() {
+void RangedClass::PrintStats() {
     std::cout << "===========STATS===========" << std::endl;
-    std::cout << "      = Melee Class =" << std::endl;
+    std::cout << "      = Ranged Class =" << std::endl;
     std::cout << "= Name: " << this -> name << std::endl;
     std::cout << "= Level: " << this -> level << std::endl;
     std::cout << "= Exp to next level: " << this -> expNext << std::endl;
